Travis CI: Workarounds for bugs in chrome / chromedriver
The previous attempt in c57f244f0adbea953dae8c8eb009c0444150dd53 was not enough
Even browser.init() fails from time to time (something goes wrong in chromedriver). However everything is fine if we retry the init call. The downside is that the test runner keeps running forever, so we need to add --exit to mocha